requirements.txt
flask
openpyxl
google-genai
python-dotenv
pandas
